distinguish between cli and cli2
