+++ aerich/utils.py
@@ -1,7 +1,8 @@
 import importlib
-from typer import Context
+
+from asyncclick import BadOptionUsage, Context
 from tortoise import BaseDBAsyncClient, Tortoise
-import typer
+
 
 def get_app_connection_name(config, app) -> str:
     """
@@ -11,16 +12,6 @@ def get_app_connection_name(config, app) -> str:
     :return:
     """
     return config.get("apps").get(app).get("default_connection", "default")
-
-
-def get_app_connection(config, app) -> BaseDBAsyncClient:
-    """
-    get connection name
-    :param config:
-    :param app:
-    :return:
-    """
-    return Tortoise.get_connection(get_app_connection_name(config, app))
 
 
 def get_tortoise_config(ctx: Context, tortoise_orm: str) -> dict:
@@ -36,21 +27,15 @@ def get_tortoise_config(ctx: Context, tortoise_orm: str) -> dict:
     try:
         config_module = importlib.import_module(config_path)
     except (ModuleNotFoundError, AttributeError):
-        typer.echo(f'No config named "{config_path}"')
-        raise typer.Exit()
+        raise BadOptionUsage(
+            ctx=ctx, message=f'No config named "{config_path}"', option_name="--config"
+        )
 
     config = getattr(config_module, tortoise_config, None)
     if not config:
-        typer.echo(f'Can\'t get "{tortoise_config}" from module "{config_module}"',)
-        raise typer.Exit()
+        raise BadOptionUsage(
+            option_name="--config",
+            message=f'Can\'t get "{tortoise_config}" from module "{config_module}"',
+            ctx=ctx,
+        )
     return config
-
-def ask_rename_column(old_name:str,new_name:str)->bool:
-    while True:
-        flag = input(f"Do you rename {old_name} to {new_name}?(y/n)")
-        if flag.lower() == "y":
-            return True
-        elif flag.lower()=="n":
-            return False
-        else:
-            print("Please input y or n.")
